Revert compact() back to an array filter()
ESLint pointed out the redundant double negation to me when I tried to
update it to the latest version. Looking at the commit that introduced
this, it seems as though I wrote this function to appease Flow. However,
now it seems that Flow is able to understand the array filter so we can
go back to using it.