Remove deprecated local configuration
This has been deprecated in favor of using methods in your configuration
file. We expect our next release to be a major release, so now is the
right time to remove this.

With this change, I had to remove a couple of tests (get aliases and get
namedExports) that were testing some details of how Configuration can be
merged, because we no longer support your configuration being an array
of objects, and we don't have any environments that do this, so this is
difficult to test, and likely not being used. I figure we can bring it
back when we need it again.