Added tests for EsModuleModule
Added tests that attempt to verify that EsModuleModule identifies every
form of export or import allowed by the current ES7 draft and the stage
1 proposals for export extensions.

The testing identified a number of problems in EsModuleModule (which
had never actually been run prior to this effort). All problems
identified have been fixed and all tests pass.

Coverage is as high as I could get it. Most of the remaining branches
are cases that should not occur, but feel right to leave. One is to
keep flow from complaining about a variable that might be null (but
actually won't). There may be some philosophy points to hash out with
these. Others are just notes at the top of the module and on an
import that I don't understand but believe have to do with post
transpilation ES5 boilerplate.