Fix imports for anonymous files
This is a quick-fix for an issue we've seen with the sublime plugin. It
would fail (with no way to recover) whenever (fix) importing to an
anonymous file.

I made an attempt at making this more general by using normalizePath to
normalize the currentFilePath but ran into a few (correctly) failing
test cases that I didn't have time to investigate proper fixes for at
the moment.