Properly resolve file paths for named exports
We've had a report in emacs-import-js [1] that `goto` wouldn't work for
named exports.

Given the following structure

  .importjs.js
  app/
    └── MyComponent.js
    └── constants/
      └── app_constants.js

the following config

  module.exports = {
    namedExports: {
      'app/constants/app_constants': [
        'Actions',
      ],
    },
    lookupPaths: [
      'app',
    ],
  }

and app/MyComponent.js containing

  import { Actions } from 'constants/app_constants';

Using `goto` on "Actions" opens `app/constants/app_constants` (which is
not a thing). It should be `app/constants/app_constants.js`.

PR 11 [1] for emacs-import-js attempts to work around this problem
locally for emacs. But this is something that applies to all import-js
plugins, so here's a fix in the core project instead.

[1]: https://github.com/Galooshi/emacs-import-js/pull/11