Use babylon to find imports
This is the first commit in a series of changes replacing our custom
regexp-powered parser with babylon [1]. I'm starting off with converting
`findCurrentImports`.

There were two tests validating that whitespace was stripped from the
top of the file. It was hard to preserve this bit of functionality so I
simply removed support for that.

I used https://astexplorer.net/ to test what the parser gave us back.

There are other parsers we could use, but I picked the one that was most
closely tied to babel, since that's what most people are using as the
standard implementation nowadays. I haven't checked performance between
different parsers, but I expect it to be good enough.

I will continue converting more of import-js to babylon. Next up is
probably going to be finding undefined/unused imports (replacing eslint
here).

[1]: https://www.npmjs.com/package/babylon