Add "star" type import to test suite
Now that we use a real es6 parser (babylon), supporting imports of the
following kind is easy:

  import * as actions from './actions';

This addresses parts of #353, but not the whole thing. We still need to
figure out when to construct those import lines. Having them be
recognized as proper imports is at least a good first step.