Replace eslint with ast parsing
Now that we have the content parsed, and methods to ask about undefined
and used identifiers, we can put that to use in place of eslint.

The eslint dependency has caused a number of annoying issues over time.
Version mismatches between the eslint used by importjs and your project
led to failing imports. We were also messing with the NODE_PATH in order
to get eslint to pick up plugins from the right place. Not great.

The only downside (afaict) of using the ast directly here is that we
won't have access to the `globals` defined through eslint. I haven't
solved that in this commit, but I expect to come back to it before we
ship this in a new release. Probably by adding a `globals` configuration
option.

It looks like this commit shaves off quite a bit of time from
fixImports. It could be that we are too naive, and we'll have to change
things later on. But hopefully it will stay faster going forward.