Remove leading ./ from import path for non-relative imports
An effect of 145d0cda was that non-relative paths were being normalized
to always start with `./`. This made imports break when
`useRelativePaths` was false.

This is a quick-fix. I spent some time trying to break out import path
creation from JsModule but failed. I hope to get back to this soon
though, because it isn't ideal.