Remove reference to lookup paths in code comment
I first wanted to remove some code here as I thought it was only valid
when multiple lookup paths were used. But after looking at the tests
around this I realized we were also solving the case when a file is
matched once through the file name and once through a package.json
pointing to the file. So the code needs to be kept, we should just
remove stuff from the comment.