Fix grabbing parentPid in daemon.js
We weren't using the right method to grab the --parent-pid option from
the command. It's likely that commander at some point changed this from
being statically accessed to being injected into the action method as an
option.

With this bug in the wild, it's more likely that people have orphaned
importjsd processes running.