Run paths through `moduleNameFormatter` when matching for goto
I found a bug when trying to goto a module that had multiple options in
the brigade codebase. We use a moduleNameFormatter to strip out parts of
the path, and for that reason, goto wasn't finding the right existing
import. By making sure that we compare paths constructed the same way,
we can bring back the correct behavior.