Keep up-to-date index of files and exports
This is still in the early stages, but the idea seems to work: Listen to
watchman changes to files, parse them for named exports, and store an
index of all named exports. Importing then becomes a quick check in this
index.

For now, I've kept the old file-scanning finders intact. I've opened up
a discussion around discontinuing these:
https://github.com/Galooshi/import-js/issues/357