Fix skipping over comments and 'use strict'
I found a bug when I tried to fix imports for a file that looked like
this:

  const Foo = Bar.Foo;

  const styles = require('index.scss');

import-js was skipping over the first statement, then removing it as
part of rewriting imports.

It turns out that babylon will ignore comments and 'use strict' anyway,
so the fix was simply to remove a condition from an if-statement,
allowing the imports to end as soon as a non-import line is found.