Switch to sqlite for exports cache
This is no longer just a cache, so I plan on renaming the module soon.

Instead of recomputing named exports for a file every single time we
start importjs, we can keep a filesystem cache of exports. I choose
sqlite here because

 - I'm comfortable with SQL
 - It will handle concurrency

On each detected file change, we first check the modification time of
the file to decide whether to update the cache/db or not.