Fix checking thousands of file in `needsUpdate()`
I ran into an issue when trying the new sqlite backed exports cache on
the brigade codebase. It has 1400+ javascript file, and with that many
files to check, the query to check mtimes ended up having too many
arguments and sqlite threw an error. I fixed this by batching the query
if needed.