Switch regular file finder over to ModuleFinder as well
In the past, we've had a few file finders available, each with
different capabilities:

 - the "find" finder - fast but mac/linux only
 - the "node" finder - cross-platform but slightly slower
 - the "watchman" finder - depending on watchman running

I've moved over watchman to ModuleFinder, and until now kept the old
ones intact. That's changing now however. On each import, we will
`ensureUpToDate()` on ModuleFinder. If the watchman watcher isn't
running, we fallback to a full filesystem scan followed by an update to
storage. This way we only have one source of truth (the db).

I had to bring back the lastUpdate module to get modification times for
files.