Don't normalize named exports
We don't want an export named e.g. _internal to match a variable named
"Internal". To make this happen, I changed the data model a little to
not store both key and name for an export. The key was the normalized
version of the name. We can instead store the raw export names for named
exports, and normalized names for default exports (since these are
constructed using the file path).