Fix findUndefinedIdentifiers for certain object references
My naive assumption about keys holding enough information to determine
whether an Identifier node is an assigment or a reference has proven to
be too naive. I found an example where we wouldn't find `uuid` as a
undefined variable in this setup:

    foo({
      bar() {},
      baz: 12,
      uuid: uuid.v4(),
    });

By passing in a context containing parent nodes (simplified) to the
visitor, we can better determine what type of identifier we're dealing
with.

I'm considering bringing in something like babel-traverse to simplify
traversing over the ast nodes. It would probably make the code more
robust. At this time however, I'm still learning the whole process of
traversing over nodes. I'd like to get some more experience on that
before I decide if it's worth bringing in a heavy-weight tool for the
job.