Prevent starting ModuleFinder in "bad" directories
Both @lencioni and I have been running into issues with watchman
starting a watch for the home directory. This folder is too big for
import-js to handle, so we shouldn't worry about it.

I decided to use a heuristic to figure out if a folder should be
watched. If either .importjs.js or package.json is present, the
directory is good. If not, we show an error.