Follow exported requires for CommonJS modules
A common thing amongst 3rd party packages is to have a main file that
looks something like this:

  module.exports = require('someOtherFile.js');

An example of this is React:

  'use strict';

  module.exports = require('./lib/React');

In order to find named exports for these types of modules, we need to
open the file being referenced and parse it for exports.

This change will set us up for reading named exports from package
dependencies as well.