Find exports from package dependencies as well
This is still work-in-progress. I want to make this more robust by
adding tests and run it more in the wild. I also need to exclude
devDependencies in some cases and deal with cleanup when dependencies
are removed.

Until now, we've only been finding exports from files outside of
node_modules. We know what dependencies a project has from its
package.json, and we can use that to get a list of node_modules files to
also parse for exports. For now, I've decided to keep the db structure
of ExportsStorage the same, and instead construct the JsModule
differently if the path starts with ./node_modules. I might revisit this
at some point if it proves easier to maintain the alias/module name
throughout the process.