Find tricky inner module.exports
I found a dependency, logpath, that was defining the default export
inside an anonymous function, in an if-statement. In order to find this
export, we would either have to traverse through all ast nodes in the
file, or run the code. I prefer not to do any of that (finding exports
needs to be fast) so I'm falling back to a regexp instead.