Don't make named exports relative
Now that we have a mechanism to automatically find named exports, the
need for namedExports configuration is not as important. One tricky
thing about namedExports is that we want to make them relative to the
current file, if they need to be. I'm removing that in favor of relying
on regular auto-named exports.