Update tests to be green
I've been ignoring test failures for a while in an attempt to move
faster. It's time to update them again so that we have a clean build.

I'm not happy about how inconsistent `goto` is about relative vs
absolute paths. I hope to address that later on, but at least this
works.