Fix resolving named imports
If two or more modules matches a variable name, you are asked to
resolve it manually. In case you selected an import that was meant to be
a named import (import { named } from 'module'), we would not honor it
being named, and the import ended up being

  import named from 'module';

I fixed this by asking the module finder once more for the import. This
is now fast since we have the storage cache/index.