Find exports for underscore.js
Because of how their exported module is constructed, finding named
exports turned out to be a little bit of a challenge. But through a
combination of remembering the defined objects, plus using a regexp to
find the actual `module.exports`, I was able to get it working.

Because this ended up being a little brittle, I added a regression test
for a bug that I was close to introducing.