Show message if cache isn't ready
Now that we initialize storage for a working directory lazily, we've
surfaced (an already existing but not before very visible) race
condition: the exports storage isn't yet populated when you first
import. The scenario is:

- You install import-js for the first time, or update an old version.
- You go to a file and run "fixImports" (or some other command)
- A background thread will start scanning the working directory for
imports. This might take a while (5-10 seconds on the Brigade codebase).
- The result comes back empty the first time.
- You wait for a bit and try again.
- Result is good.

The law of leaky abstractions [1] strikes again. In our effort to hide
complexity (finding exports in a large file system) we've reached a
point where it's tricky to hide the abstraction.

I discussed this with @lencioni, and we came up with a few ideas:
A) Have clients tell import-js about new working directories so that we
can prime the cache.
B) Let the command (fixImports, import, goto) wait until the cache is
ready.
C) Display a message that results are inaccurate.

C is the cheapest one, and that's what I've implemented in this commit.
A is expensive since it means updating all the clients. B is hard to
implement, and may lead to a pretty bad user experience for first-time
use. For instance, the vim plugin will timeout itself after 3s I
believe (built-into the channel infrastructure in vim). I'll explore A
going forward, if this quick-fix turns out to be too naive.

[1] https://www.joelonsoftware.com/2002/11/11/the-law-of-leaky-abstractions/