Find exports in files reassigning `exports`
When working on logging improvements, I added winston as a dependency.
It has a somewhat non-standard way of setting up its imports. In short:

  var winston = exports;
  exports.something = ...;

etc.

By memoizing all the aliases used for `exports`, we can catch exports of
this kind as well.

Also, since it's highly likely that people want to use the default
import from modules like this, we're also assuming a default export in
this situation.