Find exports from json files
Up until now, we haven't been treating json files differently from
js/jsx files. Which meant that babylon failed to parse the data.

By checking if the filename is .json, we can fix this by taking a
shortcut and use JSON.parse directly instead of babylon. It could be
argued that we shouldn't rely on the filename like this, but I think
it's relatively safe. If we find an example that breaks this assumption,
we can address the problem.