Fix polling-based watcher
I found a bug with how we keep ModuleFinder up-to-date when watchman
isn't available. We were applying the workingDirectory twice, leading to
corrupt paths being used. The fix includes making lastUpdate take a
workingDirectory along with the path so that it can resolve the promise
with a relative path (which we need on the receiving end).

Related to https://github.com/Galooshi/atom-import-js/issues/11