Fix support for namespace imports
When running fix imports on something that looked like this

  import * as foo from 'foo';

...we would incorrectly remove the import. This was because we weren't
properly identifying the "as foo" part as an assignment.

Since I usually don't use imports of this type, I worry that we'll
regress in the future. Even though the fix could have been isolated to
findCurrentImports, I decided to flesh out a few other tests just to
make sure we have better coverage for the fix.

Fixes #369. Thanks to @asfktz for the report!