Pick first match for goto
When using `goto` and the word couldn't be matched to a single file, we
would send back a list of `unresolvedImports` to the editor plugin.
However, none of the editor plugins (sublime, vim, emacs, atom) knows
how to resolve ambiguities for goto. In most cases, the goto path can be
resolved using current imports. In the rare event that it can't be
resolved, we now just pick the first one in the list.

This is a quick-fix to a problem that could be address more
holistically. But the more involved fix would include updating all the
plugins, and that's something I'd like to avoid right now.