Prevent findExports from blowing up on object spread syntax
I have some files that use object spread, e.g.:

```js
const foo = { ...bar, baz: true };
```

In this scenario, `key` is undefined. To prevent this from blowing up,
we need to make sure that `key` exists before trying to access its
`name` property.

There might be an opportunity to improve export detection to follow
object spreads, but that's out of the scope of this bugfix.

Fixes #373.