Exclude modules at import-time
Before, we used the `excludes` config in the ModuleFinder and Watcher
so that excluded files weren't part of the ExportsStorage at all. This
keeps the size of the cache down, but it has the annoying downside of
making the cache stale if the `excludes` config was changed by the user.

Moving to a post-storage filtering approach means that more files will
be processed by ModuleFinder, and more entries will be cached. But it
makes changes to the `excludes` config have effect right away.

In case the user is excluding a significantly large tree of files, this
commit will mean that the first cache population will take longer. I
hope this is an okay trade-off. Luckily, `node_modules` are already
handled separately/excluded which means that most projects won't notice
any change in performance.

Fixes #375