Use right mtime for node_modules dependencies
When we encounter a package.json file in the ModuleFinder, we look at
the dependencies listed and grab the exports from files inside
node_modules. An important part of this is grabbing the last modified
(mtime) timestamp, to decide whether or not to update ExportsStorage. Up
until now, we've used the mtime of the package.json file. This has led
to exports not being updated correctly as the result of an `npm install`
for instance, since package.json isn't updated when running this
command.

I'm still looking for more robust solutions to handling updates to
node_modules dependencies. The current solution requires a restart which
isn't great. I hope to have a better solution soon.