Resolve aliases that look like variables
In the Meteor environment, we have a moduleNameFormatter that will add a
slash at the beginning of paths that

- aren't a node module
- do not start with a dot

For package dependency aliases (e.g. $ => jquery) this ended up messing
up the moduleName ("/jquery"). By using the alias in a lookup from the
ModuleFinder, we make sure that by the time we hit moduleNameFormatter,
we have enough information (i.e. pathToImportedModule) to avoid adding
the slash.

Fixes #381