Delay visiting `body` arrays in findUndefinedIdentifiers
When a variable is defined after it being used inside a function scope,
`findUndefinedIdentifiers` would incorrectly flag the variable as
undefined. Here's an example:

  function foo() {
    console.log(bar);
  }

  const bar = 'fake news';

As you can see, `bar` is defined after the function making use of it.
This is perfectly valid javascript, and executing `foo()` would log
`"fake news"` to the console. To take this scenario into account, we had
to delay visiting nodes inside a function body until the parent scope
was "done" (meaning we've found all defined identifiers in it).
Essentially, this meant switching from a depth-first tree traversal to a
breadth-first. I couldn't think of a better way to do that than by
switching from the recursive version that we had before into an iterativ
solution.

This fix ended up being strongly TDD'ed. I'm not completely happy with
where the code ended up, but for now I can't come up with a more elegant
solution.

Fixes #383