Find named exports from webpack bundle
We have an application where we have a "client-core" package exporting a
bunch of reusable React components. We bundle these components using
webpack before importing them in other packages.

Before this change, the variables we exported as named exports wouldn't
be found from the webpack bundle. After, they're all correctly
identified.

There are certain things in the exported webpack file that are hard to
keep out of the named exports array. Lines like these:

exports.__esModule = true;

I decided to not worry about them, and just let them be included in the
result. It's unlikely that anyone will try to import something with that
name anyway.