Upgrade jest
I was trying to run jest in watch mode but couldn't get it to work. I
thought that an upgrade would work but it turns out I was running the
wrong `npm run` script ("test" instead of "jest"). I'm committing the
change anyway since it seems like a good idea to be on a later version.