Improve recursive export finding
Now that we have require-relative, we can make use of it when finding
exports recursively in a file. Before, we would just assume that the
relatively `require`d module was located in a nested node_modules
folder. Now we find ones that have been hoisted to the topmost
node_modules folder as well, along with ones that are implicitly
depended on from elsewhere.