Gracefully handle imports with local names for import
Import-js never produces import statements with local names to named
imports, e.g.

  import { foo as bar } from 'car';

But there's no guarantee that the file is free from them, so we need to
handle them gracefully. I did that by making the `namedImports` array
contain objects with the local and (optionally) imported name. On import
statement production, we then use the imported name if we need to.

Fixes #407