Configure Babylon to understand dynamic imports
Now that dynamic imports are a stage 3 proposal [0], more people are
going to start using this syntax and we should support it out of the
box. Without this plugin, import-js will crash when run on a file that
uses this syntax.

[0]: https://github.com/tc39/proposal-dynamic-import

We may eventually want to expose the array of Babylon plugins so they
can be configured by the end user, but I think we should wait to do that
until we actually see the need for it.

Addresses #428