Replace lodash per-method packages with deep imports
The per-method packages are no longer maintained, and since they are
dependency-free they often end up bringing in more code than they
necessarily need to. This aspect does not matter much for this project,
but it seems like a good idea to get on the regular lodash package so we
can get updates going forward.