Teach `findUndefinedIdentifiers` about hoisting
Fixes #415

Because of how js assignments are hoisted to the top of its scope, we
need to wait with deciding if a variable is undefined until the entire
scope is processed.

I thought about making a two-pass implementation here, but decided it's
easier to just hold on to the current context and apply a
post-processing filter at the very end.

The new implementation is a little heavier on memory usage (holding on
to all identifiers + contexts until the end). I tested this on the
largest file I can think of in this project (the Importer-test.js file)
and didn't notice any degradation.