Make `goto` find more file paths
Before when running `goto`, we would ask the ModuleFinder for matches on
the variable name. We would then take that list and map it against
current imports. This only worked in cases where the file you were going
to was already in the index. When the file for some reason wasn't in the
index, goto wouldn't work.

To make things a little more robust, I'm changing the ordering of things
a bit. Instead of asking the index first, we look through current
imports and try to resolve the module name for the matching import
statement. If that fails (the variable isn't yet imported or the path
couldn't be resolved), we continue to ask the ModuleFinder for matches.

The new behavior is a little bit more robust, and slightly faster.

The changes made in #457 inspired me to dig a little deeper through
`goto` which led up to this commit.