Avoid using Array.prototype.includes
We want to support node versions down to 4. As a result of that, we need
to be a little careful about what newer ES features we use.
Array.prototype.includes is one of them (Object.assign another).

I first went down the path of finding a polyfill to use instead of
resorting to indexOf. There's babel-plugin-array-includes [1], which
transforms code from .includes to .indexOf, but it only works for things
that can be statically determined as arrays. There's the babel-polyfill
[2] which would work but comes at a significant weight. There's also the
option of hand-picking a polyfill from core-js [3], but core-js is also
a quite large dependency which I'd prefer to avoid. Introducing
polyfills could also come back and bite us when the Atom plugin is
updated. This plugin uses import-js as a library (and not as a
standalone daemon like the other plugins).

We didn't catch this in the PR introducing the `.includes` becase Travis
wasn't configured to run on PRs. I've since changed that, so any use of
unsupported methods in the future should get caught on the PR level.

[1] https://www.npmjs.com/package/babel-plugin-array-includes
[2] https://babeljs.io/docs/usage/polyfill/
[3] https://github.com/zloirock/core-js#commonjs