Remove dependency on `deasync`
We used this because it took too long to convert our old sync tests to
promises when we made the whole lib async. Since we were using node 6
and 7 we couldn't use async/await directly so we turned to deasync. Now
that 6 and 7 are gone, we can convert to async/await.

This fixes a hard-to-debug build failure in Travis where the test run
would just stall on Node 12 and 13.