Disable codeclimate when env var is missing
When outside contributors, e.g. dependabot, pushes a PR, the
CODECLIMATE_REPO_TOKEN env variable is not present (it's a secret
defined on the github account). We can ignore the codeclimate report in
these cases to let tests still pass.

An alternative option would be to make the codeclimate secret available
to everyone, but that's not very safe.

Fixes #619