Remove flow type annotations
We know that we do not want to use flow anymore. We may want to adopt
TypeScript, but it is unclear if we will in the near term.

https://github.com/Galooshi/import-js/issues/593

I want to re-introduce ESLint, and instead of setting up ESLint to be
able to parse flow syntax, I decided that we should just remove it.

https://github.com/Galooshi/import-js/issues/608

I removed these annotations by running the following command:

```
npx flow-remove-types lib/ -d lib/
```

Then I formatted the files by running prettier on everything:

```
node_modules/.bin/prettier -w lib/
```

I think if we wanted to introduce TypeScript that it might be easier to
do it from scratch anyway instead of attempting to convert the flow
types over to a different type system.